Refactor nodeValues
