+++ packages/frontend/editor-ui/src/utils/nodeSettingsUtils.test.ts
@@ -6,6 +6,7 @@ import type {
 	IDataObject,
 	INodeTypeDescription,
 	INodePropertyOptions,
+	INodeParameters,
 } from 'n8n-workflow';
 import {
 	updateDynamicConnections,
@@ -13,9 +14,11 @@ import {
 	nameIsParameter,
 	formatAsExpression,
 	parseFromExpression,
+	setValue,
 } from './nodeSettingsUtils';
 import { SWITCH_NODE_TYPE } from '@/constants';
 import type { INodeUi, IUpdateInformation } from '@/Interface';
+import { type Ref, ref } from 'vue';
 
 describe('updateDynamicConnections', () => {
 	afterAll(() => {
@@ -394,5 +397,49 @@ describe('parseFromExpression', () => {
 
 	it('returns null for other types if value is undefined', () => {
 		expect(parseFromExpression({}, undefined, 'json', null, [])).toBeNull();
+	});
+});
+
+describe('setValue', () => {
+	let nodeValues: Ref<INodeParameters>;
+	beforeEach(() => {
+		nodeValues = ref({
+			color: '#ff0000',
+			alwaysOutputData: false,
+			executeOnce: false,
+			notesInFlow: false,
+			onError: 'stopWorkflow',
+			retryOnFail: false,
+			maxTries: 3,
+			waitBetweenTries: 1000,
+			notes: '',
+			parameters: {},
+		});
+	});
+
+	it('mutates nodeValues as expected', () => {
+		setValue(nodeValues, 'color', '#ffffff');
+
+		expect(nodeValues.value.color).toBe('#ffffff');
+		expect(nodeValues.value.parameters).toEqual({});
+
+		setValue(nodeValues, 'parameters.key', 3);
+
+		expect(nodeValues.value.parameters).toEqual({ key: 3 });
+
+		nodeValues.value = { parameters: { some: { nested: {} } } };
+		setValue(nodeValues, 'parameters.some.nested.key', true);
+
+		expect(nodeValues.value.parameters).toEqual({
+			some: { nested: { key: true } },
+		});
+
+		setValue(nodeValues, 'parameters', null);
+
+		expect(nodeValues.value.parameters).toBe(undefined);
+
+		setValue(nodeValues, 'newProperty', 'newValue');
+
+		expect(nodeValues.value.newProperty).toBe('newValue');
 	});
 });
